server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==0.28.1
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
httpx==0.25.2
google-auth-oauthlib==1.1.0
google-auth==2.23.4
google-api-python-client==2.108.0
PyJWT==2.8.0
google-auth-httplib2==0.2.0
google-generativeai>=0.7.2
sqlalchemy==1.4.53
psycopg2-binary==2.9.9
alembic==1.13.0
asyncpg==0.29.0
databases[postgresql]==0.8.0
apscheduler==3.10.4
python-dateutil==2.8.2
pytz>=2021.1
# Security Dependencies (removed cryptography - not needed for simple token refresh)
# Content Analysis Dependencies
pandas>=2.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
html5lib>=1.1
